requirements.txt
# Core LangChain dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.40

# Streamlit for chat interface
streamlit>=1.28.0

# LangSmith for tracing and observability
langsmith>=0.0.80

# Environment variable management
python-dotenv>=1.0.0
